analyzer task: move settle planning and anime preview derivation out of the season loop
SeasonReanalysisPlanner now owns every settle-reanalysis decision (mode selection, derived-mode expansion, the episode-set comparison that lived in AnalysisHelpers). Anime preview derivation moves to Analyzers/AnimePreviewDeriver. The analyzer chain is built from three optional instances and one preferred instance instead of a switch over predicates; the two one-line helpers and their tests are inlined. Drops the plugin-null guard, the two log-and-rethrow catches and five one-use debug stubs.
