+++ IntroSkipper/Data/AnalysisHelpers.cs
@@ -31,20 +31,6 @@ internal static class AnalysisHelpers
         ModeToSegmentType.ToDictionary(pair => pair.Value.ToString(), pair => pair.Key, StringComparer.OrdinalIgnoreCase);
 
     /// <summary>
-    /// Returns whether a settled-season analysis mode still needs re-analysis for its current episode
-    /// set. Pure set comparison: the decision is committed separately via
-    /// <see cref="Db.IIntroSkipperDatabase.RecordSettleReanalysisAsync(Guid, IReadOnlyCollection{AnalysisMode}, IReadOnlyCollection{Guid}, CancellationToken)"/>
-    /// once the reset has succeeded, so the completed episode set survives plugin restarts.
-    /// </summary>
-    /// <param name="settledEpisodeIds">Episode IDs recorded when the season was last settle-reanalyzed for this mode.</param>
-    /// <param name="episodeIds">Current episode IDs in the season.</param>
-    /// <returns><see langword="true"/> when a re-analysis should be performed; otherwise <see langword="false"/>.</returns>
-    internal static bool ShouldSettleReanalyze(
-        IReadOnlySet<Guid> settledEpisodeIds,
-        IReadOnlyCollection<Guid> episodeIds)
-        => settledEpisodeIds.Count != episodeIds.Count || episodeIds.Any(id => !settledEpisodeIds.Contains(id));
-
-    /// <summary>
     /// Returns whether the mode has a <see cref="ModeToSegmentType"/> entry, i.e. every
     /// downstream conversion of a row carrying it is defined. Write boundaries and HTTP
     /// edges reject modes failing this so no stored row can crash a later mirror; a mere
